Update PyPI version when Github release is created
Added new Github Actions workflow to auto-update source code with the release tag as the package version.

Already generated and configured the package-specific PyPI API token as a repository secret.

Tested using the existing unit tests and the equivalent process being executed for the CLI repo. Will not be able to test fully end-to-end until a new release is published.